+++ transcendence-api/src/auth/strategy/jwt.strategy.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import { UserService } from '../../user/user.service';
 
-import { TokenPayload } from '../auth.model';
+import { UserService } from '../../user/user.service';
+import { TokenPayload } from '../utils/auth.model';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
