+++ transcendence-api/src/auth/utils/auth.model.ts
